In courses application, move helpers from views to a utils module.
